Advisors

+++ UtilityFunctions.cc
@@ -367,7 +367,7 @@ bool hasPrefix (string prefix, string candidate) {
 bool yearMonthDay (const string& date, int& year, int& month, int& day) {
   int firstDotIdx = -1;
   int secondDotIdx = -1;
-  for (int i = 0; i < date.size(); ++i) {
+  for (unsigned int i = 0; i < date.size(); ++i) {
     if (isdigit(date[i])) continue;
     if (date[i] != '.') return false;
     if (firstDotIdx < 0) firstDotIdx = i;
@@ -377,7 +377,7 @@ bool yearMonthDay (const string& date, int& year, int& month, int& day) {
   if (firstDotIdx < 0) return false;
   if (secondDotIdx < 0) return false;
   if (firstDotIdx + 1 == secondDotIdx) return false;
-  if (secondDotIdx + 1 == date.size()) return false;
+  if (secondDotIdx + 1 == (int) date.size()) return false;
   year = atoi(date.substr(0, firstDotIdx).c_str());
   month = atoi(date.substr(firstDotIdx+1, secondDotIdx).c_str());
   day = atoi(date.substr(secondDotIdx).c_str());
